Removed unused imports
fixed warnings

+++ src/main/java/me/toast/clicks/Utils.java
@@ -22,7 +22,7 @@ public class Utils {
     }
     public static Color RainbowEffect(long offset, float brightness, int speed) {
         float hue = (float) (System.nanoTime() + (offset * speed)) / 1.0E10F % 1.0F;
-        long color = Long.parseLong(Integer.toHexString(Integer.valueOf(Color.HSBtoRGB(hue, brightness, 1F)).intValue()), 16);
+        long color = Long.parseLong(Integer.toHexString(Color.HSBtoRGB(hue, brightness, 1F)), 16);
         Color c = new Color((int) color);
         return new Color(c.getRed()/255.0F, c.getGreen()/255.0F, c.getBlue()/255.0F, c.getAlpha()/255.0F);
     }
